requirements.txt
torch==2.8.0
torchvision==0.23.0
pillow==11.3.0
py-real-esrgan==2.0.0
huggingface-hub==0.23.0
<<<<<<< HEAD
flask==3.0.3
=======
>>>>>>> 3c57f4522786e0671a18d30f5778165443c6f258
